+++ packages/ui/src/api/stravaActivities/queryStravaActivities.ts
@@ -5,7 +5,7 @@ import { API_ENDPOINTS } from '../constants';
 
 /**
  * Queries Strava activities.
- * @returns {Promise<StravaActivity[]>} Strava activities.
+ * @returns {Promise<StravaActivity[] | null>} Strava activities.
  */
 const queryStravaActivities = (): Promise<StravaActivity[] | null> =>
   client<StravaActivity[]>(API_ENDPOINTS.STRAVA_ACTIVITIES);