+++ packages/ui/src/api/stravaActivities/useQueryStravaActivities.ts
@@ -1,4 +1,5 @@
 import { useQuery } from '@tanstack/react-query';
+import { StravaActivity } from '@torq/strava-api';
 
 import { API_ENDPOINTS } from '../constants';
 
@@ -7,11 +8,11 @@ import { API_ENDPOINTS } from '../constants';
  * @returns {object} Strava activities data.
  */
 const useQueryStravaActivities = () =>
-  useQuery({
+  useQuery<StravaActivity[] | null>({
     queryKey: [API_ENDPOINTS.STRAVA_ACTIVITIES],
     /**
      * Queries Strava activities from the internal API.
-     * @returns {Promise<StravaActivity[]>} Strava activities.
+     * @returns {Promise<StravaActivity[] | null>} Strava activities.
      */
     queryFn: async () => {
       const { default: queryStravaActivities } = await import('./queryStravaActivities');