+++ packages/ui/src/api/stravaActivity/queryStravaActivity.ts
@@ -6,9 +6,9 @@ import client from '../client';
 /**
  * Query specific activity by ID.
  * @param {string} id - Activity ID
- * @returns {Promise<StravaActivity>} Activity data
+ * @returns {Promise<StravaActivity | null>} Activity data
  */
 const queryStravaActivity = (id: string): Promise<StravaActivity | null> =>
-  client<StravaActivity>(API_ENDPOINTS.STRAVA_ACTIVITY(id));
+  client<StravaActivity | null>(API_ENDPOINTS.STRAVA_ACTIVITY(id));
 
 export default queryStravaActivity;